Add upgrade infra
