+++ .env.example
@@ -26,4 +26,16 @@ ROPSTEN_PRIVATE_KEY=
 # from: https://docs.chain.link/docs/ethereum-addresses#config
 CHAINLINK_MAINNET_PRICE_FEED=0xdeb288F737066589598e9214E782fa5A8eD689e8
 CHAINLINK_KOVAN_PRICE_FEED=0xF7904a295A029a3aBDFFB6F12755974a958C7C25
+CHAINLINK_GOERLI_PRICE_FEED='There isnt one'
 CHAINLINK_RINKEBY_PRICE_FEED=0x2431452A0010a43878bF198e170F6319Af6d27F4
+CHAINLINK_ROPSTEN_PRICE_FEED='There isnt one'
+
+# These are the proxy addresses of the permanant testnet deployments
+# useful for testing upgrades across environments
+RAREPIZZAS_BOX_MAINNET_PROXY_ADDRESS=
+RAREPIZZAS_BOX_GOERLI_PROXY_ADDRESS=
+RAREPIZZAS_BOX_RINKEBY_PROXY_ADDRESS=
+
+# Proxy Administrator
+RAREPIZZAS_BOX_MAINNET_PROXY_ADMIN_ADDRESS=0xBA5E28a2D1C8cF67Ac9E0dfc850DC8b7b21A4DE2
+RAREPIZZAS_BOX_RINKEBY_PROXY_ADMIN_ADDRESS=0x465cDAC68140B62c11a1E89eCD30e97aA5a18413
