+++ .gitignore
@@ -23,4 +23,4 @@ node_modules
 npm-debug.log*
 
 # Openzeppelin
-.openzeppelin
+.openzeppelin/unknown-*.json
