+++ README.md
@@ -29,7 +29,7 @@ npx hardhat node
 - deploy contracts locally (in a different terminal)
 
 ```
-npx hardhat run --network localhost scripts/deploy-box.proxy.ts
+npx hardhat run --network localhost scripts/box.deploy.ts
 ```
 
 - deploy contracts remotely
@@ -37,7 +37,7 @@ npx hardhat run --network localhost scripts/deploy-box.proxy.ts
 copy .env.example to .env and fill in your keys
 
 ```
-$ npx hardhat run --network goerli scripts/deploy-box.proxy.ts
+$ npx hardhat run --network rinkeby scripts/box.deploy.ts
 ```
 
 then go to etherscan and get the implementation address (under the contract source, theres a `more options` dropdown menu and select `is this a proxy?`)
@@ -59,3 +59,30 @@ npx hardhat test
 ```
 npx hardhat test --watch
 ```
+
+## Upgrading
+
+Basically do this:
+
+https://forum.openzeppelin.com/t/openzeppelin-upgrades-step-by-step-tutorial-for-hardhat/3580
+
+
+```bash
+# transfer ownership if you havent already
+npx hardhat run --network rinkeby scripts/box.transfer.proxy.ts
+```
+
+```bash
+# deploy the upgraded contract with your current signing key
+npx hardhat run --network rinkeby scripts/box.upgrade.V2.prepare.ts
+```
+
+then, go to gnosis and open the openzepplin app and put in the proxy and new implementation and send a request
+
+
+## dev
+
+we use goerli as a developer test network and we use rinkeby for UAT.
+
+opensea doesnt currently work with goerli
+neither does gnosis
