+++ config/index.ts
@@ -16,6 +16,11 @@ const config: NetworkConfig = {
   CHAINLINK_GOERLI_PRICE_FEED: process.env.CHAINLINK_GOERLI_PRICE_FEED || '0x0000000000000000000000000000000000000000',
   CHAINLINK_RINKEBY_PRICE_FEED: process.env.CHAINLINK_RINKEBY_PRICE_FEED || '0x0000000000000000000000000000000000000000',
   CHAINLINK_ROPSTEN_PRICE_FEED: process.env.CHAINLINK_ROPSTEN_PRICE_FEED || '0x0000000000000000000000000000000000000000',
+  RAREPIZZAS_BOX_MAINNET_PROXY_ADDRESS: process.env.RAREPIZZAS_BOX_MAINNET_PROXY_ADDRESS || '',
+  RAREPIZZAS_BOX_GOERLI_PROXY_ADDRESS: process.env.RAREPIZZAS_BOX_GOERLI_PROXY_ADDRESS || '',
+  RAREPIZZAS_BOX_RINKEBY_PROXY_ADDRESS: process.env.RAREPIZZAS_BOX_RINKEBY_PROXY_ADDRESS || '',
+  RAREPIZZAS_BOX_MAINNET_PROXY_ADMIN_ADDRESS: process.env.RAREPIZZAS_BOX_MAINNET_PROXY_ADMIN_ADDRESS || '',
+  RAREPIZZAS_BOX_RINKEBY_PROXY_ADMIN_ADDRESS: process.env.RAREPIZZAS_BOX_RINKEBY_PROXY_ADMIN_ADDRESS || '',
 }
 
 export default config
